Cleaned workspace+Added working lstm
