+++ FakeNewsDetection/articlePrediction/urls.py
@@ -2,6 +2,6 @@ from unicodedata import name
 from django.urls import path,include
 from . import views
 urlpatterns = [
-    path('',views.home,name='home'),
-    path('result/', views.index, name='index'),
+    path('', views.index, name='index'),
+    path('lstm/',views.home,name='home'),
 ]
